Fix scanning of strings and implement \z
The old pattern (in the 3rd alternative) allowed \ to be scanned alone
and not only as the first char in a pair which meant that the string
ending '/" could become quoted by mistake. We also now read in the
whole string and then work out if any included eol is correct.