Make the random state part of the Lua state
Taken ideas from @bopjesvla and extended them to handle the older
random module as well. Need to be cunning here and take code from the
older random module as it did not have the necessary seed
function. This IS safe as the module has been unchanged for many
versions and will disappear soon.

This fixes #78.
This closes #84.