Started working on simplifying stack and uvpvalue handling
Not done everywhere yet so not everything works yet.

We will now no longer optimise the stacks as it is cheap to have
unused frames in them. We just add them as 'not_used'. This will
simplify the handling which will make it easier to trace call stacks.

We also snuck in using longer record and key names to make them more
inderstandable.