Put the heap in a separate record included in the luerl state
This make the luerl state record much smaller which will Hopefully
make it more efficient when updating state and not cost much more when
updating the heap. An alternative could be to put set of 3 elements
per table into a separate record.

Initial teswts show the speed about the same in our benchmarks.