We now converted to a flat emulator loop
The emulator has a gflat loop stepping over instructions so function
calls and loops are explicitly kept in a call stack so the Erlang call
stack is no longerused for this. Some instructions do get a bit
complex and for loops the instruction list is dynamically managed. One
of the issues with immutable memory.

This restructuring will make it easier to move towards using binaries
as code if we ever want to go that way.