Move low level heap functions out of luerl_emul into luerl_heap
These functions now only work on the actual heap and do not call emul
functions but return values saying when this has to be done. The
library modules have been updated to use the new module.

We also took the chance to update the copyright years on the affected
modules.