Add a {file,FileName} option to the compiler
This allows the user to define which file a chunk looks like it has
come from. This can be useful when loading code.

We have also taken the chancew to redo the interfaces to the compiler
modules called from the compiler. The new API is better. This only
affects the modules' interface functions and not their internals.